Removed Mr. Savage's Speech Bubble
I couldn't have Mr. Savage be dynamically sized and still have the text bubble and the text on it line up right so I had to settle a blurred box off to the side. I also added a class for the blurred boxes so they don't all have to have seperate styles.